requirements.txt
Django==3.0
psycopg2-binary==2.8.6
amqp==2.6.1
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.0
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0
django-extra-fields==3.0.2
django-filter==2.4.0
django-phone-field==1.8.1
django-templated-mail==1.1.1
djangorestframework==3.10.0
djangorestframework-gis==0.16
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.5
drf-extra-fields==3.0.2
drf-yasg==1.17.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
kombu==4.6.11
MarkupSafe==1.1.1
packaging==20.4
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2020.1
requests==2.24.0
ruamel.yaml==0.16.12
redis
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
vine==1.3.0
